Use Decimal for calculations

+++ models/input.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from pydantic.dataclasses import dataclass
 from enum import Enum
+from decimal import Decimal
 
 
 class Action(Enum):
@@ -20,21 +21,21 @@ class Event:
     isin: str = None
     ticker: str = None
     name: str = None
-    shares_count: float = 0
-    price_per_share: float = 0
+    shares_count: Decimal = 0
+    price_per_share: Decimal = 0
     price_currency: str = None
-    exchange_rate: float = 0
-    result_eur: float = 0
-    total_eur: float = 0
-    withholding_tax: float = 0
+    exchange_rate: Decimal = 0
+    result_eur: Decimal = 0
+    total_eur: Decimal = 0
+    withholding_tax: Decimal = 0
     withholding_tax_currency: str = None
-    charge_amount_eur: float = 0
-    deposit_fee_eur: float = 0
-    stamp_duty_tax_eur: float = 0
+    charge_amount_eur: Decimal = 0
+    deposit_fee_eur: Decimal = 0
+    stamp_duty_tax_eur: Decimal = 0
     notes: str = None
     ID: str = None
-    french_tax: float = 0
-    conversion_fee_eur: float = 0
+    french_tax: Decimal = 0
+    conversion_fee_eur: Decimal = 0
 
     class Config:
         anystr_strip_whitespace = True
